Assets
Sorry for the delay again!
Here are the rest of the SciFi items in a single sprite sheet: All icons
are 256x256px.
The arrows for directions for the kitchen and scifi levels. also
256x256px.
ALSO finally remembered to upload the bread. Sorry this is all so
last-minute. I was hit by some severe stomach pains and put on pain meds
that made me a little drowsy